- some refactoring
