+++ app/src/main/java/com/racovita/wow/data/models/Product.kt
@@ -1,11 +1,12 @@
 package com.racovita.wow.data.models
 
-import androidx.room.*
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.racovita.wow.data.db.FAVORITE_PRODUCTS
 
 /**
  * Used in adapter & DB
  */
-
 @Entity(tableName = FAVORITE_PRODUCTS)
 data class Product(
     @PrimaryKey val id: Int,
