+++ app/src/main/java/com/racovita/wow/features/base/repo/BaseRepo.kt
@@ -3,6 +3,9 @@ package com.racovita.wow.features.base.repo
 import com.racovita.wow.data.db.AppDatabase
 import com.racovita.wow.data.models.Product
 
+/**
+ * Common functional for all repos
+ */
 open class BaseRepo(
     val database: AppDatabase
 ) {
