+++ app/src/main/java/com/racovita/wow/utils/UtilsDi.kt
@@ -6,7 +6,6 @@ import org.koin.dsl.module
 import org.koin.experimental.builder.single
 
 val utilsModule = module {
-
     single<ResUtil>()
     single<RestClient>()
 }
