+++ app/src/main/res/layout/content_details.xml
@@ -4,7 +4,7 @@
     android:id="@+id/zone_details"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.racovita.wow.features.details.view.ActivityDetails"
+    tools:context="com.racovita.wow.features.details.view.DetailsActivity"
     tools:showIn="@layout/activity_details">
 
     <LinearLayout
@@ -111,7 +111,7 @@
 
         <TextView
             android:id="@+id/tv_description"
-            style="@style/TextView.Secondary"
+            style="@style/TextView.Primary"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="@dimen/material_margin"
